rebird the option to disable deletes when not required anymore

+++ src/main/java/com/purbon/kafka/topology/TopicManager.java
@@ -1,12 +1,16 @@
 package com.purbon.kafka.topology;
 
+import static com.purbon.kafka.topology.BuilderCLI.ALLOW_DELETE_OPTION;
+
 import com.purbon.kafka.topology.model.Project;
 import com.purbon.kafka.topology.model.Topic;
 import com.purbon.kafka.topology.model.Topology;
 import java.io.PrintStream;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -19,9 +23,17 @@ public class TopicManager {
   public static final String REPLICATION_FACTOR = "replication.factor";
 
   private final TopologyBuilderAdminClient adminClient;
+  private final Map<String, String> cliParams;
+  private final Boolean allowDelete;
 
   public TopicManager(TopologyBuilderAdminClient adminClient) {
+    this(adminClient, new HashMap<>());
+  }
+
+  public TopicManager(TopologyBuilderAdminClient adminClient, Map<String, String> cliParams) {
     this.adminClient = adminClient;
+    this.cliParams = cliParams;
+    this.allowDelete = Boolean.valueOf(cliParams.getOrDefault(ALLOW_DELETE_OPTION, "true"));
   }
 
   public void sync(Topology topology) {
@@ -44,17 +56,19 @@ public class TopicManager {
                           updatedListOfTopics.add(fullTopicName);
                         }));
 
-    // Handle topic delete: Topics in the initial list, but not present anymore after a
-    // full topic sync should be deleted
-    List<String> topicsToBeDeleted = new ArrayList<>();
-    listOfTopics.stream()
-        .forEach(
-            originalTopic -> {
-              if (!updatedListOfTopics.contains(originalTopic)) {
-                topicsToBeDeleted.add(originalTopic);
-              }
-            });
-    adminClient.deleteTopics(topicsToBeDeleted);
+    if (allowDelete) {
+      // Handle topic delete: Topics in the initial list, but not present anymore after a
+      // full topic sync should be deleted
+      List<String> topicsToBeDeleted = new ArrayList<>();
+      listOfTopics.stream()
+          .forEach(
+              originalTopic -> {
+                if (!updatedListOfTopics.contains(originalTopic)) {
+                  topicsToBeDeleted.add(originalTopic);
+                }
+              });
+      adminClient.deleteTopics(topicsToBeDeleted);
+    }
   }
 
   public void syncTopic(Topic topic, String fullTopicName, Set<String> listOfTopics) {
